Move FieldRef into CIRCT core
Moving FieldRef into CIRCT core, and separating the FIRRTL specific
functionality into new functions.
